==> notebooks/3_model_reg.ipynb <==
# Abalone Project - Model Regression
Maria Eugênia Fonseca\
2021/09/28

import pandas as pd
import matplotlib as plt
import seaborn as sns
import mlflow
from sklearn import linear_model
from sklearn.ensemble import GradientBoostingRegressor
from sklearn.model_selection import RandomizedSearchCV, GridSearchCV
from sklearn.metrics import mean_squared_error, r2_score

X_train = pd.read_csv("../data/processed/abalone_xtrain.csv")
y_train = pd.read_csv("../data/processed/abalone_ytrain.csv")

X_test = pd.read_csv("../data/processed/abalone_xtest.csv")
y_test = pd.read_csv("../data/processed/abalone_ytest.csv")

X_train.head()

#### Start mlflow experiment:

EXPERIMENT_NAME = "abalone"

try:
    mlflow.create_experiment(EXPERIMENT_NAME)
    print("CREATING")
except:
    print("ALREADY EXISTS")

mlflow.set_experiment(EXPERIMENT_NAME)

#### Function to evaluate regression and log metrics in mlflow:

def evaluate_regression_and_log(regression_model, X_train, y_train, X_test, y_test, params):    
    # metrics - train
    y_train_pred = regression_model.predict(X_train)
    rmse_train = mean_squared_error(y_train, y_train_pred, squared=False)
    r2_train = r2_score(y_train, y_train_pred)

    # metrics - test
    y_test_pred = regression_model.predict(X_test)
    rmse_test = mean_squared_error(y_test, y_test_pred, squared=False)
    r2_test = r2_score(y_test, y_test_pred)
    
    # log metrics    
    mlflow.start_run()

    mlflow.log_metric("rmse_train", rmse_train)
    mlflow.log_metric("r2_train", r2_train)
    mlflow.log_metric("rmse_test", rmse_test)
    mlflow.log_metric("r2_test", r2_test)

    mlflow.sklearn.log_model(regression_model, "model")
    mlflow.log_param("model_name", type(regression_model).__name__)
    mlflow.log_params(params)
    
    mlflow.end_run()

params_pipe = {
    "boxcox_transformer": "false",
    "minmax_scaller": "true"}

#### First model - LinearRegression

lm = linear_model.LinearRegression()
lm = lm.fit(X_train, y_train)

evaluate_regression_and_log(lm, X_train, y_train, X_test, y_test, params_pipe)

#### Second model - GradientBoostingRegressor default

params = {
    "boxcox_transformer": "false",
    "minmax_scaller": "true",
    "hyperparameter_tuning": "false"}

gbr_base = GradientBoostingRegressor(random_state=123)
gbr_base = gbr_base.fit(X_train, y_train)

evaluate_regression_and_log(gbr_base, X_train, y_train, X_test, y_test, params)

gbr_base.get_params()

#### Third model - GradientBoostingRegressor with hyperparameter tuning

params = {
    "boxcox_transformer": "false",
    "minmax_scaller": "true",
    "hyperparameter_tuning": "true"}

learning_rate = [0.03, 0.04, 0.05, 0.06]
n_estimators = [140, 150, 160]
subsample = [0.7, 0.75, 0.8, 0.85]
min_samples_leaf = [10, 12, 15]
max_depth = [4, 5, 6] 
max_features = ['auto', 'sqrt']
warm_start = [True]

# Create the grid
random_grid = {'learning_rate': learning_rate,
               'n_estimators': n_estimators,
               'subsample': subsample,
               'max_features': max_features,
               'min_samples_leaf': min_samples_leaf,
               'max_depth': max_depth,
               'max_features': max_features,
               'warm_start': warm_start}

gbr = GradientBoostingRegressor(random_state=123)

# Random search of parameters, using 3 fold cross validation, 
gbr_tuning = GridSearchCV(estimator=gbr, param_grid=random_grid, cv=3, verbose=10, n_jobs=-1)

# Fit the random search model
gbr_tuning.fit(X_train, y_train)

gbr_tuning.best_params_

gbr_tuned = gbr_tuning.best_estimator_

evaluate_regression_and_log(gbr_tuned, X_train, y_train, X_test, y_test, params)

feat_imp_gbr = {'column_name': X_train.columns, 
            'importance': gbr_tuned.feature_importances_}

feat_imp_gbr = pd.DataFrame(feat_imp_gbr).sort_values(by=['importance'], ascending=False)

f, ax = plt.subplots(figsize=(10, 8))

sns.barplot(x=feat_imp_gbr.iloc['importance'], y=feat_imp_gbr.iloc['column_name'], color='#94d1e3')

ax.text(x=0, y=1, s='Feature importance', fontsize=18, weight='bold', ha='left', va='bottom', transform=ax.transAxes)

ax.set(xlabel="Importance", ylabel="")
plt.xticks(np.arange(0, max(feat_imp_gbr[:20]['importance']), 0.1))

ax.grid(False)
sns.despine(bottom = False, left = False)
sns.set_style("whitegrid") 

#plt.savefig('/content/drive/My Drive/Maria Eugênia Fonseca/Gráficos e mapas/top20_gbr_comodidades.png', quality=100)
plt.show()

#### Forth model - Lasso default
