==> notebooks/5_model_classification.ipynb <==
# Abalone Project - Model Classification
Maria Eugênia Fonseca\
2021/10/05

import pandas as pd
import mlflow

from sklearn.metrics import accuracy_score, f1_score
from sklearn.naive_bayes import CategoricalNB
from sklearn.neighbors import KNeighborsClassifier
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
from sklearn.model_selection import GridSearchCV

X_train = pd.read_csv("../data/processed/abalone_xtrain_class.csv")
y_train = pd.read_csv("../data/processed/abalone_ytrain_class.csv")

X_test = pd.read_csv("../data/processed/abalone_xtest_class.csv")
y_test = pd.read_csv("../data/processed/abalone_ytest_class.csv")

X_train.head()

y_train.head()

#### Start mlflow experiment:

EXPERIMENT_NAME = "abalone_classification"

try:
    mlflow.create_experiment(EXPERIMENT_NAME)
    print("CREATING")
except:
    print("ALREADY EXISTS")

mlflow.set_experiment(EXPERIMENT_NAME)

#### Function to evaluate classification and log metrics in mlflow:

def evaluate_classification_and_log(classification_model, X_train, y_train, X_test, y_test, params):    
    # metrics - train
    y_train_pred = classification_model.predict(X_train)
    accuracy_train = accuracy_score(y_train, y_train_pred)
    f1_train = f1_score(y_train, y_train_pred, average='weighted')

    # metrics - test
    y_test_pred = classification_model.predict(X_test)
    accuracy_test = accuracy_score(y_test, y_test_pred)
    f1_test = f1_score(y_test, y_test_pred, average='weighted')
    
    # log metrics    
    mlflow.start_run()

    mlflow.log_metric("accuracy_train", accuracy_train)
    mlflow.log_metric("f1_train", f1_train)
    mlflow.log_metric("accuracy_test", accuracy_test)
    mlflow.log_metric("f1_test", f1_test)

    mlflow.sklearn.log_model(classification_model, "model")
    mlflow.log_param("model_name", type(classification_model).__name__)
    mlflow.log_params(params)
    
    mlflow.end_run()

#### First model - CategoricalNB

params_pipe = {
    "minmax_scaller": "true"}

nb = CategoricalNB().fit(X_train, y_train)

evaluate_classification_and_log(nb, X_train, y_train, X_test, y_test, params_pipe)

#### Second model - KNeighborsClassifier

knn = KNeighborsClassifier().fit(X_train, y_train)

evaluate_classification_and_log(knn, X_train, y_train, X_test, y_test, params_pipe)

knn.get_params()

#### Third model - KNeighborsClassifier with hyperparameter tuning

params_pipe = {
    "minmax_scaller": "true",
    "hyperparameter_tuning": "true"}

n_neighbors = [4, 5, 6, 7, 8]
weights = ['uniform', 'distance']

# Create the grid
search_grid = {'n_neighbors': n_neighbors,
               'weights': weights}

knn = KNeighborsClassifier()

# Grid search of parameters, using 3 fold cross validation, 
knn_tuning = GridSearchCV(estimator=knn, param_grid=search_grid, cv=3, n_jobs=-1)

# Fit the random search model
knn_tuning.fit(X_train, y_train)

knn_tuning.best_params_

knn_tuned = knn_tuning.best_estimator_

evaluate_classification_and_log(knn_tuned, X_train, y_train, X_test, y_test, params_pipe)

#### Fourth model - RandomForestClassifier no tuning

params_pipe = {
    "minmax_scaller": "true",
    "hyperparameter_tuning": "false"}

rf = RandomForestClassifier(random_state=123).fit(X_train, y_train)

evaluate_classification_and_log(rf, X_train, y_train, X_test, y_test, params_pipe)

rf.get_params()

#### Fith model - RandomForestClassifier with hyperparameter tuning

params_pipe = {
    "minmax_scaller": "true",
    "hyperparameter_tuning": "true"}

n_estimators = [115, 120, 125]
max_depth = [6, 7, 8]
min_samples_leaf = [6, 7, 8]
ccp_alpha = [0.0009, 0.001, 0.003]
max_samples = [0.68, 0.7, 0.72]

# Create the grid
search_grid = {'n_estimators': n_estimators,
               'max_depth': max_depth,
               'min_samples_leaf': min_samples_leaf,
               'ccp_alpha': ccp_alpha,
               'max_samples': max_samples
              }

rf = RandomForestClassifier(random_state=123)

# Grid search of parameters, using 3 fold cross validation, 
rf_tuning = GridSearchCV(estimator=rf, param_grid=search_grid, cv=3, n_jobs=-1)

# Fit the random search model
rf_tuning.fit(X_train, y_train)

rf_tuning.best_params_

rf_tuned = rf_tuning.best_estimator_

evaluate_classification_and_log(rf_tuned, X_train, y_train, X_test, y_test, params_pipe)

#### Sixth model - GradientBoostingClassifier with hyperparameter tuning

params_pipe = {
    "minmax_scaller": "true",
    "hyperparameter_tuning": "false"}

gb = GradientBoostingClassifier(random_state=123).fit(X_train, y_train)

evaluate_classification_and_log(gb, X_train, y_train, X_test, y_test, params_pipe)

#### Seventh model - GradientBoostingClassifier with hyperparameter tuning

params_pipe = {
    "minmax_scaller": "true",
    "hyperparameter_tuning": "true"}

learning_rate = [0.05, 0.1]
n_estimators = [100, 120, 140]
subsample = [0.6, 0.7, 0.8]
min_samples_leaf = [6, 7, 8]
max_depth = [6, 7, 8]
ccp_alpha = [0.0001, 0.001, 0.01]

# Create the grid
search_grid = {'learning_rate': learning_rate,
               'n_estimators': n_estimators,
               'subsample': subsample,
               'min_samples_leaf': min_samples_leaf,
               'max_depth': max_depth,
               'ccp_alpha': ccp_alpha
              }

gb = GradientBoostingClassifier(random_state=123)

# Grid search of parameters, using 3 fold cross validation, 
gb_tuning = GridSearchCV(estimator=gb, param_grid=search_grid, cv=3, n_jobs=-1)

# Fit the random search model
gb_tuning.fit(X_train, y_train)

gb_tuning.

gb_tuned = gb_tuning.best_estimator_

evaluate_classification_and_log(gb_tuned, X_train, y_train, X_test, y_test, params_pipe)
