==> notebooks/6_creating_train_and_predict_samples.ipynb <==
# Abalone Project - Model Classification
Maria Eugênia Fonseca\
2022/02/22

Separating the original dataset into samples to train and to predict.

import pandas as pd
from sklearn.model_selection import train_test_split
import numpy as np

df_raw_data = pd.read_csv("../data/raw/abalone_data.txt", header=None)

train, predict = train_test_split(df_raw_data, test_size=0.2, random_state=123)

# removing target variable from predict
predict = predict.iloc[: , :-1]

train.to_csv("../data/raw/train.txt", header=None, index=None, sep=',')

predict.to_csv("../data/raw/to_predict.txt", header=None, index=None, sep=',')
